requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
sentence-transformers
langdetect
transformers
numpy
pandas
scikit-learn
nltk
elasticsearch
kafka-python
pyyaml
